move copy button into its own component
