+++ src/components/DonationOption.svelte
@@ -3,13 +3,7 @@
 	export let text;
 	export let showQrToggle;
 
-	let copied = false;
-
-	const copy = (field) => {
-		navigator.clipboard.writeText(field);
-		copied = true;
-		setTimeout(() => (copied = false), 2100);
-	};
+	import { CopyButton } from '$comp';
 </script>
 
 <div>
@@ -55,29 +49,7 @@
 			</button>
 
 			<!-- copy button -->
-			<button class="text-link hover:text-hover transition-colors" on:click={() => copy(value)}>
-				{#if copied}
-					<svg
-						xmlns="http://www.w3.org/2000/svg"
-						viewBox="0 0 512 512"
-						fill="currentColor"
-						class="w-6 h-6"
-						><!--! Font Awesome Pro 6.2.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc. --><path
-							d="M470.6 105.4c12.5 12.5 12.5 32.8 0 45.3l-256 256c-12.5 12.5-32.8 12.5-45.3 0l-128-128c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L192 338.7 425.4 105.4c12.5-12.5 32.8-12.5 45.3 0z"
-						/></svg
-					>
-				{:else}
-					<svg
-						xmlns="http://www.w3.org/2000/svg"
-						viewBox="0 0 384 512"
-						fill="currentColor"
-						class="w-6 h-6"
-						><!--! Font Awesome Pro 6.2.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc. --><path
-							d="M192 0c-41.8 0-77.4 26.7-90.5 64H48C21.5 64 0 85.5 0 112V464c0 26.5 21.5 48 48 48H336c26.5 0 48-21.5 48-48V112c0-26.5-21.5-48-48-48H282.5C269.4 26.7 233.8 0 192 0zm0 128c-17.7 0-32-14.3-32-32s14.3-32 32-32s32 14.3 32 32s-14.3 32-32 32zm-80 64H272c8.8 0 16 7.2 16 16s-7.2 16-16 16H112c-8.8 0-16-7.2-16-16s7.2-16 16-16z"
-						/></svg
-					>
-				{/if}
-			</button>
+			<CopyButton {value} />
 		</div>
 	</div>
 </div>
